# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FastAccess accessible stadium navigation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type VenueNodeType = typeof VenueNodeType[keyof typeof VenueNodeType];


export const VenueNodeType = {
  gate: 'gate',
  junction: 'junction',
  section: 'section',
  ramp: 'ramp',
  elevator: 'elevator',
  quiet_room: 'quiet_room',
  sensory_area: 'sensory_area',
  restroom: 'restroom',
  service_animal_relief: 'service_animal_relief',
  guest_services: 'guest_services',
  first_aid: 'first_aid',
} as const;

export interface VenueNode {
  id: string;
  label: string;
  type: VenueNodeType;
}

export type RouteStepAccessType = typeof RouteStepAccessType[keyof typeof RouteStepAccessType];


export const RouteStepAccessType = {
  'step-free': 'step-free',
  ramp: 'ramp',
  elevator: 'elevator',
  'stairs-only': 'stairs-only',
} as const;

export type RouteStepNoiseLevel = typeof RouteStepNoiseLevel[keyof typeof RouteStepNoiseLevel];


export const RouteStepNoiseLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export type RouteStepLighting = typeof RouteStepLighting[keyof typeof RouteStepLighting];


export const RouteStepLighting = {
  steady: 'steady',
  flashing: 'flashing',
} as const;

export interface RouteStep {
  node: VenueNode;
  distanceM: number;
  cumulativeM: number;
  accessType: RouteStepAccessType;
  noiseLevel: RouteStepNoiseLevel;
  lighting: RouteStepLighting;
}

export interface RouteResult {
  steps: RouteStep[];
  totalDistanceM: number;
  profiles: string[];
  found?: boolean;
}

export type SensoryEventType = typeof SensoryEventType[keyof typeof SensoryEventType];


export const SensoryEventType = {
  noise: 'noise',
  light: 'light',
  both: 'both',
} as const;

export type SensoryEventIntensity = typeof SensoryEventIntensity[keyof typeof SensoryEventIntensity];


export const SensoryEventIntensity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface SensoryEvent {
  minute: number;
  name: string;
  type: SensoryEventType;
  intensity: SensoryEventIntensity;
  /** @nullable */
  durationMinutes?: number | null;
}

export interface QueryInput {
  text: string;
  /** @nullable */
  context?: string | null;
}

export interface QueryResponse {
  echo: string;
  stub: boolean;
  message?: string;
}

export interface DescribeInput {
  nodeId: string;
  profiles?: string[];
}

export interface DescribeResponse {
  nodeId: string;
  description: string;
  stub: boolean;
}

export type GetRouteParams = {
from: string;
to: string;
profiles?: string[];
};

export type GetForecastParams = {
minute: number;
window?: number;
};