# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FastAccess accessible stadium navigation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /route:
    get:
      operationId: getRoute
      tags: [route]
      summary: Get accessible route between two venue nodes
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
        - name: to
          in: query
          required: true
          schema:
            type: string
        - name: profiles
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
      responses:
        "200":
          description: Computed accessible route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResult"
        "400":
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No route found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    VenueNode:
      type: object
      required: [id, label, type]
      properties:
        id:
          type: string
        label:
          type: string
        type:
          type: string
          enum:
            - gate
            - junction
            - section
            - ramp
            - elevator
            - quiet_room
            - sensory_area
            - restroom
            - service_animal_relief
            - guest_services
            - first_aid
    RouteStep:
      type: object
      required: [node, distanceM, cumulativeM, accessType, noiseLevel, lighting]
      properties:
        node:
          $ref: "#/components/schemas/VenueNode"
        distanceM:
          type: number
        cumulativeM:
          type: number
        accessType:
          type: string
          enum: [step-free, ramp, elevator, stairs-only]
        noiseLevel:
          type: string
          enum: [low, medium, high]
        lighting:
          type: string
          enum: [steady, flashing]
    RouteResult:
      type: object
      required: [steps, totalDistanceM, profiles]
      properties:
        steps:
          type: array
          items:
            $ref: "#/components/schemas/RouteStep"
        totalDistanceM:
          type: number
        profiles:
          type: array
          items:
            type: string
        found:
          type: boolean
...