# lib/api-zod/src/generated/types/routeResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FastAccess accessible stadium navigation API
 * OpenAPI spec version: 0.1.0
 */
import type { RouteStep } from './routeStep';

export interface RouteResult {
  steps: RouteStep[];
  totalDistanceM: number;
  profiles: string[];
  found?: boolean;
}