# lib/api-zod/src/generated/types/routeStep.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FastAccess accessible stadium navigation API
 * OpenAPI spec version: 0.1.0
 */
import type { RouteStepAccessType } from './routeStepAccessType';
import type { RouteStepLighting } from './routeStepLighting';
import type { RouteStepNoiseLevel } from './routeStepNoiseLevel';
import type { VenueNode } from './venueNode';

export interface RouteStep {
  node: VenueNode;
  distanceM: number;
  cumulativeM: number;
  accessType: RouteStepAccessType;
  noiseLevel: RouteStepNoiseLevel;
  lighting: RouteStepLighting;
}